day 248: change all types to int64 for portability

+++ day248/problem.go
@@ -1,7 +1,7 @@
 package day248
 
 // MaxIf is a typical max function using an if branch.
-func MaxIf(a, b int) int {
+func MaxIf(a, b int64) int64 {
 	if a > b {
 		return a
 	}
@@ -10,6 +10,6 @@ func MaxIf(a, b int) int {
 
 // MaxSubtractAndShift finds the max using subtraction
 // and right shifting.
-func MaxSubtractAndShift(a, b int) int {
+func MaxSubtractAndShift(a, b int64) int64 {
 	return a - ((a - b) & ((a - b) >> 63))
 }
